fixed bug in NodeData. Should not properly choose between x86 and x64 versions of dll

+++ MotionFieldsTest/Assets/Scripts/SO_MotionField.cs
@@ -101,9 +101,9 @@ namespace AnimationMotionFields {
 
 	public class NodeData{
 		public string clipId;
-		public int timeSamp;
+		public float timeSamp;
 
-		public NodeData(string id, int time){
+		public NodeData(string id, float time){
 			this.clipId = id;
 			this.timeSamp = time;
 		}
